feat: 初始化 car_recommendation 项目 -更新新闻资讯
