+++ src/main/webapp/newshow.jsp
@@ -8,29 +8,38 @@
     <style>
         body {
             font-family: 'Arial', sans-serif;
-            background-color: #f4f4f4;
+            background-color: #f0f0f0; /* 浅灰色背景 */
             margin: 0;
             padding: 0;
+            color: #333; /* 默认字体颜色 */
         }
         .container {
             width: 80%;
             margin: 20px auto;
             padding: 20px;
-        }
-        .news-card {
-            background-color: #fff;
+            background-color: #ffffff; /* 白色背景 */
             border-radius: 8px;
             box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+        }
+        .news-card {
             margin-bottom: 20px;
             padding: 15px;
             display: flex;
             flex-direction: row;
             align-items: center;
+            background-color: #eef2f5; /* 浅蓝色背景 */
+            border-radius: 8px;
+            box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+            transition: transform 0.3s ease-in-out; /* 平滑过渡效果 */
+        }
+        .news-card:hover {
+            transform: translateY(-5px); /* 鼠标悬停时上移 */
         }
         .news-card img {
-            width: 150px; /* Adjust as needed */
+            width: 150px;
             height: auto;
             margin-right: 15px;
+            border-radius: 4px; /* 图片圆角 */
         }
         .news-card div {
             flex-grow: 1;
@@ -38,13 +47,15 @@
         .news-title {
             font-size: 1.2em;
             margin: 0 0 10px 0;
+            color: #2a4055; /* 标题颜色 */
         }
         .news-summary {
-            color: #666;
-            text-decoration: none; /* Remove underline from the link */
+            color: #007bff; /* 链接颜色 */
+            text-decoration: none;
         }
         .news-summary:hover {
-            text-decoration: underline; /* Add underline on hover */
+            text-decoration: underline;
+            color: #0056b3; /* 悬停时链接颜色 */
         }
     </style>
 </head>
